Deny cast_possible_truncation to avoid expensive cast checking in GPU code (#101)
* examples/matrix: Avoid expensive usize -> u32 conversion

* Fix tests

* gpu/chunk_impl: Avoid expensive downcast

* Improve code to avoid cast_possible_truncation

* CI: deny clippy::cast_possible_truncation

* emit metadata

+++ crates/rustc_codegen_gpu/tests/codegen_fails/diversed_chunk2.rs
@@ -8,14 +8,14 @@
 #[gpu::device]
 #[no_mangle]
 #[inline(never)]
-fn dev_call(a: u32, b: u32) -> u32 {
-    (a + b) + 1
+fn dev_call(a: u32, b: u32) -> usize {
+    ((a + b) + 1) as usize
 }
 
 #[gpu::device]
 #[inline(always)]
-fn dev_call_inline(a: u32, b: u32) -> u32 {
-    (a + b) + 1
+fn dev_call_inline(a: u32, b: u32) -> usize {
+    ((a + b) + 1) as usize
 }
